fix segfault, set mCurrentFrame.mTcw back to empty if reloc failed
During the relocalization function (starts on line 1343), we set the
mCurrentFrame.mTcw to an estimated pose and then try to refine it.
But if we can't refine it enough, then we exit the relocalization function
without emptying the mTcw pose matrix on the current frame. Then, later in
Track(..) we try to save the the current pose in order to reconstruct the
trajectory later. In order to do this we check if the current frame has a pose,
and since it is not empty because of the checks during reloc, we try to add a
non-existent reference keyframe. So the solution is in relocalization to set
the mTcw matrix back to empty before exiting reloc if we fail to relocalize.
by @bmende